# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crucible API Service (Microservices Mode)",
    "description": "API service for distributed Crucible platform",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/eval": {
      "post": {
        "summary": "Evaluate",
        "description": "Submit code for evaluation",
        "operationId": "evaluate_api_eval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/eval-status/{eval_id}": {
      "get": {
        "summary": "Get Evaluation Status",
        "description": "Get evaluation status from storage",
        "operationId": "get_evaluation_status_api_eval_status__eval_id__get",
        "parameters": [
          {
            "name": "eval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Eval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations": {
      "get": {
        "summary": "Get Evaluations",
        "description": "Get evaluation history",
        "operationId": "get_evaluations_api_evaluations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "python"
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "default": "docker"
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "EvaluationRequest"
      },
      "EvaluationResponse": {
        "properties": {
          "eval_id": {
            "type": "string",
            "title": "Eval Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "queued"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Evaluation queued for processing"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          }
        },
        "type": "object",
        "required": [
          "eval_id"
        ],
        "title": "EvaluationResponse"
      },
      "EvaluationStatusResponse": {
        "properties": {
          "eval_id": {
            "type": "string",
            "title": "Eval Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "default": ""
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "eval_id",
          "status"
        ],
        "title": "EvaluationStatusResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}